Add utility files and rename arteliveitem file
- Add .kodiignore to avoid package scripts and tests
- Add __init__.py in mapper folder
- Rename live.py into arteliveitem.py for naming consistency
