+++ resources/lib/mapper/arteliveitem.py
